pyproject.toml
[tool.poetry]
name = "crypto-forecast-with-bytewax"
version = "0.1.0"
description = ""
authors = ["Osinachi Chukwujama  <Osinachi.Chukwujama@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
bytewax = "^0.9.0"
websocket-client = "^1.3.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
